Move all options to new assembly instead of linking source files.
