+++ src/Microsoft.Diagnostics.Monitoring.Options/ApiAuthenticationOptions.cs
@@ -5,11 +5,7 @@
 using Microsoft.Diagnostics.Monitoring.WebApi;
 using System.ComponentModel.DataAnnotations;
 
-#if UNITTEST
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
-#else
 namespace Microsoft.Diagnostics.Tools.Monitor
-#endif
 {
     internal sealed class ApiAuthenticationOptions
     {
