+++ src/Microsoft.Diagnostics.Monitoring.Options/DiagnosticPortOptions.cs
@@ -5,13 +5,7 @@
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
-#if UNITTEST
-using Microsoft.Diagnostics.Monitoring.WebApi;
-
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
-#else
 namespace Microsoft.Diagnostics.Monitoring.WebApi
-#endif
 {
     public class DiagnosticPortOptions
     {
