+++ src/Microsoft.Diagnostics.Monitoring.Options/IEgressProviderCommonOptions.cs
@@ -2,11 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-#if UNITTEST
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
-#else
 namespace Microsoft.Diagnostics.Tools.Monitor.Egress
-#endif
 {
     /// <summary>
     /// Egress provider options common to all egress providers.
