+++ src/Microsoft.Diagnostics.Monitoring.Options/ProcessFilterOptions.cs
@@ -7,13 +7,7 @@ using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
-#if UNITTEST
-using Microsoft.Diagnostics.Monitoring.WebApi;
-
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
-#else
 namespace Microsoft.Diagnostics.Monitoring.WebApi
-#endif
 {
     [JsonConverter(typeof(JsonStringEnumConverter))]
     public enum ProcessFilterKey
