+++ src/Microsoft.Diagnostics.Monitoring.Options/StorageOptions.cs
@@ -4,13 +4,7 @@
 
 using System.ComponentModel.DataAnnotations;
 
-#if UNITTEST
-using Microsoft.Diagnostics.Monitoring.WebApi;
-
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
-#else
 namespace Microsoft.Diagnostics.Monitoring.WebApi
-#endif
 {
     internal class StorageOptions
     {
