+++ src/Microsoft.Diagnostics.Monitoring.Options/StorageOptionsDefaults.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.IO;
+
+namespace Microsoft.Diagnostics.Monitoring.WebApi
+{
+    internal sealed class StorageOptionsDefaults
+    {
+        public static readonly string DumpTempFolder = Path.GetTempPath();
+    }
+}
