+++ src/Tests/Microsoft.Diagnostics.Monitoring.Tool.FunctionalTests/Options/RootOptions.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.Diagnostics.Monitoring.TestCommon.Options
+using Microsoft.Diagnostics.Monitoring.WebApi;
+
+namespace Microsoft.Diagnostics.Tools.Monitor
 {
     internal class RootOptions
     {
