+++ src/Tests/Microsoft.Diagnostics.Monitoring.WebApi.UnitTests/EndpointInfoSourceTests.cs
@@ -201,7 +201,7 @@ namespace Microsoft.Diagnostics.Monitoring.WebApi.UnitTests
 
         private TestServerEndpointInfoSource CreateServerSource(out string transportName)
         {
-            DiagnosticPortHelper.Generate(TestCommon.Options.DiagnosticPortConnectionMode.Listen, out _, out transportName);
+            DiagnosticPortHelper.Generate(DiagnosticPortConnectionMode.Listen, out _, out transportName);
             _outputHelper.WriteLine("Starting server endpoint info source at '" + transportName + "'.");
             return new TestServerEndpointInfoSource(transportName, _outputHelper);
         }
@@ -209,7 +209,7 @@ namespace Microsoft.Diagnostics.Monitoring.WebApi.UnitTests
         private AppRunner CreateAppRunner(string transportName = null, int appId = 1)
         {
             AppRunner appRunner = new(_outputHelper, Assembly.GetExecutingAssembly(), appId);
-            appRunner.ConnectionMode = TestCommon.Options.DiagnosticPortConnectionMode.Connect;
+            appRunner.ConnectionMode = DiagnosticPortConnectionMode.Connect;
             appRunner.DiagnosticPortPath = transportName;
             appRunner.ScenarioName = TestAppScenarios.AsyncWait.Name;
             return appRunner;
